finish the seeders and factories, and adding decumentation to methods, before that I added a Command colors to the output. this commit is also fixed the key type and key name for models

+++ src/javabaseproject/javabase/framework/FileHandler.java
@@ -1,18 +1,26 @@
 package javabaseproject.javabase.framework;
 
-import javabaseproject.javabase.framework.commandline.Command;
-import javabaseproject.javabase.framework.commandline.output.Colors;
-
 import java.io.*;
 
 public class FileHandler {
+    private static boolean dirsCreated;
     private final File file;
     public FileHandler(String filePath) {
         this(new File(filePath));
     }
 
     public FileHandler(File file){
+        this(file, true);
+    }
+
+    private FileHandler(File file, boolean makeDefaultsDirs){
         this.file = file;
+        if(makeDefaultsDirs)
+            makeDefaultsDirs();
+    }
+
+    private static FileHandler of(String file, boolean makeDefaultDirs){
+        return new FileHandler(new File(file), makeDefaultDirs);
     }
 
     public static FileHandler of(String filePath) {
@@ -29,20 +37,44 @@ public class FileHandler {
     }
 
     public String read() throws IOException {
-        String readed = "";
+        StringBuilder readed = new StringBuilder();
         BufferedReader fr = new BufferedReader(new FileReader(file));
         String temp;
         while((temp = fr.readLine()) != null){
-            readed += temp;
+            readed.append("\n").append(temp);
         }
-        return readed;
+        return readed.toString();
     }
 
     public boolean delete(){
-        return file.delete();
+        if(file.exists()){
+            return file.delete();
+        }else{
+            return false;
+        }
     }
 
     public boolean exists(){
         return file.exists();
+    }
+
+    public boolean makeDir(){
+        return file.mkdirs();
+    }
+
+    private void makeDefaultsDirs(){
+        if(dirsCreated)
+            return;
+        FileHandler fh ;
+        if(!(fh = FileHandler.of(FilePaths.getModelsPath(), false)).exists()){
+            fh.makeDir();
+        }
+        if(!(fh = FileHandler.of(FilePaths.getSeedersPath(), false)).exists()){
+            fh.makeDir();
+        }
+        if(!(fh = FileHandler.of(FilePaths.getFactoriesPath(), false)).exists()){
+            fh.makeDir();
+        }
+        dirsCreated = true;
     }
 }
